requirements.txt
anthropic
httpx
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
beautifulsoup4
lxml
sqlite-utils
fastapi
uvicorn
